Double zoom delivery - feature/#640
Changes not related to double zoom:
* CacheManager: new ITileSource member instead of MapTileProviderBase

Light changes:
* org.osmdroid.views.MapController: plus unrelated `zoomToSpan(int latSpanE6, int lonSpanE6)` rewriting
* org.osmdroid.google.wrapper.MapView
* MinimapOverlay
* MyLocationNewOverlay
* OpenStreetMapConstants: new tag in order to get/put the double zoom in prefs
* OsmdroidMapWrapper
* OsmPath
* Polyline
* Projection
* SampleOsmPath
* SimpleFastPointOverlay
* StarterMapFragment: using new tag in order to get/put the double zoom in prefs
* ZoomEvent

Tricky changes:
* IMapController: deprecated 3 integer zoom methods; created 3 double zoom methods instead
* IMapView: deprecated 1 integer zoom method; created 1 double zoom methods instead
* org.osmdroid.google.wrapper.MapController: plus unrelated `zoomToSpan(int latSpanE6, int lonSpanE6)` rewriting
* MapTileProviderBase
* org.osmdroid.views.MapView
* PathProjection: to be double-checked
* TileLooper: now we need to make the distinction between two tile sizes - for input it's basically 256, for output it's between 256 (integer zoom) and 511.99999 (almost the next integer zoom)
* TilesOverlay: handled the distinction between the two tile sizes, and the rounding issues
* org.osmdroid.util.TileSystem: deprecated some integer zoom methods, created the corresponding double zoom methods (among them `getFactor`, which computes real powers of 2 instead of << and >>), had to copy some code from microsoft.mappoint.TileSystem as we needed different tile sizes

+++ osmdroid-android/src/main/java/org/osmdroid/events/ZoomEvent.java
@@ -9,9 +9,9 @@ import org.osmdroid.views.MapView;
  */
 public class ZoomEvent implements MapEvent {
 	protected MapView source;
-	protected int zoomLevel;
+	protected double zoomLevel;
 
-	public ZoomEvent(final MapView source, final int zoomLevel) {
+	public ZoomEvent(final MapView source, final double zoomLevel) {
 		this.source = source;
 		this.zoomLevel = zoomLevel;
 	}
@@ -26,7 +26,7 @@ public class ZoomEvent implements MapEvent {
 	/*
 	 * Return the zoom level zoomed to.
 	 */
-	public int getZoomLevel() {
+	public double getZoomLevel() {
 		return zoomLevel;
 	}
